refactor(typescript): transform apidom-ns-openapi3 package
Refs #18
