+++ apidom/.editorconfig
@@ -28,6 +28,3 @@ trim_trailing_whitespace = true
 # Denotes whether file should end with a newline
 # Possible values - true, false
 insert_final_newline = true
-
-[*.ts]
-indent_size = 4
