+++ apidom/babel.config.json
@@ -1,4 +1,7 @@
 {
+  "babelrcRoots": [
+    "packages/*"
+  ],
   "env": {
     "commonjs": {
       "presets": [
