+++ apidom/packages/apidom-ns-openapi3/.gitignore
@@ -1,0 +1,3 @@
+/dist
+/es
+/lib
