+++ apidom/packages/apidom-ns-openapi3/config/webpack/browser.config.js
@@ -1,0 +1,72 @@
+'use strict';
+
+const path = require('path');
+const { nonMinimizeTrait, minimizeTrait } = require('./traits.config');
+
+const browser = {
+  mode: 'production',
+  entry: ['./src/polyfills.ts', './src/namespace.ts'],
+  target: 'web',
+  performance: {
+    maxEntrypointSize: 712000,
+    maxAssetSize: 712000,
+  },
+  output: {
+    path: path.resolve('./dist'),
+    filename: 'apidom-ns-openapi3.browser.js',
+    libraryTarget: 'umd',
+    library: 'apidomNsOpenapi3',
+  },
+  resolve: {
+    extensions: ['.ts', '.mjs', '.js', '.json'],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(ts|js)?$/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            babelrc: true,
+            rootMode: 'upward',
+          },
+        },
+      },
+    ],
+  },
+  ...nonMinimizeTrait,
+};
+
+const browserMin = {
+  mode: 'production',
+  entry: ['./src/polyfills.ts', './src/namespace.ts'],
+  target: 'web',
+  output: {
+    path: path.resolve('./dist'),
+    filename: 'apidom-ns-openapi3.browser.min.js',
+    libraryTarget: 'umd',
+    library: 'apidomNsOpenapi3',
+  },
+  resolve: {
+    extensions: ['.ts', '.mjs', '.js', '.json'],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(ts|js)?$/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            babelrc: true,
+            rootMode: 'upward',
+          },
+        },
+      },
+    ],
+  },
+  ...minimizeTrait,
+};
+
+module.exports = [browser, browserMin];
