+++ apidom/packages/apidom-ns-openapi3/config/webpack/traits.config.js
@@ -1,0 +1,39 @@
+'use strict';
+
+const webpack = require('webpack');
+const TerserPlugin = require('terser-webpack-plugin');
+
+const nonMinimizeTrait = {
+  optimization: {
+    minimize: false,
+    usedExports: false,
+    concatenateModules: false,
+  },
+};
+
+const minimizeTrait = {
+  plugins: [
+    new webpack.LoaderOptionsPlugin({
+      minimize: true,
+    }),
+  ],
+  optimization: {
+    minimizer: [
+      new TerserPlugin({
+        terserOptions: {
+          compress: {
+            warnings: false,
+          },
+          output: {
+            comments: false,
+          },
+        },
+      }),
+    ],
+  },
+};
+
+module.exports = {
+  nonMinimizeTrait,
+  minimizeTrait,
+};
