+++ apidom/packages/apidom-ns-openapi3/package.json
@@ -2,8 +2,23 @@
   "name": "apidom-ns-openapi3",
   "version": "0.1.0",
   "description": "OpenAPI3 profile for ApiDOM",
-  "main": "./src/namespace.js",
-  "scripts": {},
+  "main": "lib/namespace.js",
+  "module": "es/namespace.js",
+  "jsnext:main": "es/namespace.js",
+  "browser": "dist/apidom-ns-openapi3.browser.js",
+  "unpkg": "dist/apidom-ns-openapi3.browser.min.js",
+  "scripts": {
+    "build": "cross-env npm run build:es && npm run build:commonjs && npm run build:umd",
+    "build:es": "cross-env BABEL_ENV=es babel src --out-dir es --extensions '.ts' --root-mode 'upward'",
+    "build:commonjs": "cross-env BABEL_ENV=commonjs babel src --out-dir lib --extensions '.ts' --root-mode 'upward'",
+    "build:umd": "cross-env npm run build:umd:node && npm run build:umd:browser",
+    "build:umd:node": "cross-env BABEL_ENV=node webpack --config config/webpack/node.config.js --progress",
+    "build:umd:browser": "cross-env BABEL_ENV=browser webpack --config config/webpack/browser.config.js --progress",
+    "lint": "eslint ./",
+    "lint:fix": "eslint ./ --fix",
+    "clean": "rimraf ./es ./lib ./dist",
+    "check-types": "tsc --noEmit"
+  },
   "author": "Vladimir Gorej",
   "license": "Apache-2.0",
   "dependencies": {
