--- apidom/packages/apidom-ns-openapi3/src/elements/Components.js
@@ -1,24 +1,0 @@
-'use strict';
-
-const {ObjectElement} = require('minim');
-
-/**
- * @class Components
- *
- * @param {object} content
- * @param meta
- * @param attributes
- * @extends ObjectElement
- */
-class Components extends ObjectElement {
-  constructor(...args) {
-    super(...args);
-    this.element = 'components';
-  }
-
-  get schemas() {
-    return this.get('schemas');
-  }
-}
-
-module.exports = Components;
