+++ apidom/packages/apidom-ns-openapi3/src/elements/Contact.ts
@@ -1,0 +1,34 @@
+import { Attributes, Meta, ObjectElement } from 'minim';
+
+class Contact extends ObjectElement {
+  constructor(content: Record<string, unknown>, meta: Meta, attributes: Attributes) {
+    super(content, meta, attributes);
+    this.element = 'contact';
+  }
+
+  get name(): string {
+    return this.get('name').toValue();
+  }
+
+  set name(name: string) {
+    this.set('name', name);
+  }
+
+  get url(): string {
+    return this.get('url');
+  }
+
+  set url(url: string) {
+    this.set('url', url);
+  }
+
+  get email(): string {
+    return this.get('email');
+  }
+
+  set email(email: string) {
+    this.set('email', email);
+  }
+}
+
+export default Contact;
