--- apidom/packages/apidom-ns-openapi3/src/elements/OpenApi3.js
@@ -1,29 +1,0 @@
-'use strict';
-
-const {ObjectElement} = require('minim');
-
-/**
- * @class OpenApi3
- *
- * @param {object} content
- * @param meta
- * @param attributes
- * @extends ObjectElement
- */
-class OpenApi3 extends ObjectElement {
-  constructor(...args) {
-    super(...args);
-    this.element = 'openApi3';
-    this.classes.push('api');
-  }
-
-  get openapi() {
-    return this.get('openapi');
-  }
-
-  get info() {
-    return this.get('info');
-  }
-}
-
-module.exports = OpenApi3;
