--- apidom/packages/apidom-ns-openapi3/src/elements/Openapi.js
@@ -1,20 +1,0 @@
-'use strict';
-
-const {StringElement} = require('minim');
-
-/**
- * @class Openapi *
- * @param {string} content
- * @param meta
- * @param attributes
- *
- * @extends StringElement
- */
-class Openapi extends StringElement {
-  constructor(...args) {
-    super(...args);
-    this.element = 'openapi';
-  }
-}
-
-module.exports = Openapi;
