--- apidom/packages/apidom-ns-openapi3/src/elements/Schema.js
@@ -1,20 +1,0 @@
-'use strict';
-
-const {ObjectElement} = require('minim');
-
-/**
- * @class Schema
- *
- * @param {object} content
- * @param meta
- * @param attributes
- * @extends ObjectElement
- */
-class Schema extends ObjectElement {
-  constructor(...args) {
-    super(...args);
-    this.element = 'schema';
-  }
-}
-
-module.exports = Schema;
