+++ apidom/packages/apidom-ns-openapi3/src/polyfills.ts
@@ -1,0 +1,3 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import 'core-js/stable';
+import 'regenerator-runtime/runtime';
