+++ apidom/packages/apidom-ns-openapi3/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": [
+    "../@types/**/*.d.ts",
+    "src/**/*",
+    "test/**/*"
+  ]
+}
