--- apidom/packages/apidom/.editorconfig
@@ -1,30 +1,0 @@
-# http://editorconfig.org
-
-# A special property that should be specified at the top of the file outside of
-# any sections. Set to true to stop .editor config file search on current file
-root = true
-
-[*]
-# Indentation style
-# Possible values - tab, space
-indent_style = space
-
-# Indentation size in single-spaced characters
-# Possible values - an integer, tab
-indent_size = 2
-
-# Line ending file format
-# Possible values - lf, crlf, cr
-end_of_line = lf
-
-# File character encoding
-# Possible values - latin1, utf-8, utf-16be, utf-16le
-charset = utf-8
-
-# Denotes whether to trim whitespace at the end of lines
-# Possible values - true, false
-trim_trailing_whitespace = true
-
-# Denotes whether file should end with a newline
-# Possible values - true, false
-insert_final_newline = true
