+++ apidom/packages/apidom/package.json
@@ -9,11 +9,11 @@
   "unpkg": "dist/apidom.browser.min.js",
   "scripts": {
     "build": "cross-env npm run build:es && npm run build:commonjs && npm run build:umd",
-    "build:es": "cross-env BABEL_ENV=es babel src --out-dir es --extensions '.ts'",
-    "build:commonjs": "cross-env BABEL_ENV=commonjs babel src --out-dir lib --extensions '.ts'",
+    "build:es": "cross-env BABEL_ENV=es babel src --out-dir es --extensions '.ts' --root-mode 'upward'",
+    "build:commonjs": "cross-env BABEL_ENV=commonjs babel src --out-dir lib --extensions '.ts' --root-mode 'upward'",
     "build:umd": "cross-env npm run build:umd:node && npm run build:umd:browser",
-    "build:umd:node": "cross-env BABEL_ENV=node webpack --config config/webpack/webpack.config.js --progress",
-    "build:umd:browser": "cross-env BABEL_ENV=browser webpack --config config/webpack/webpack.config.js --progress",
+    "build:umd:node": "cross-env BABEL_ENV=node webpack --config config/webpack/node.config.js --progress",
+    "build:umd:browser": "cross-env BABEL_ENV=browser webpack --config config/webpack/browser.config.js --progress",
     "lint": "eslint ./",
     "lint:fix": "eslint ./ --fix",
     "clean": "rimraf ./es ./lib ./dist",
