+++ apidom/packages/apidom/test/mocha-bootstrap.js
@@ -1,2 +1,2 @@
 require('regenerator-runtime');
-require('@babel/register')({ extensions: ['.js', '.ts'] });
+require('@babel/register')({ extensions: ['.js', '.ts'], rootMode: 'upward' });
